+++ lib/utils/i_app_navigator.dart
@@ -1,6 +1,10 @@
 import 'package:flutter/material.dart';
 
 abstract class IAppNavigator {
+
+  GlobalKey<NavigatorState> get navigatorKey;
+
+
   Future<T?> pushNamed<T extends Object?>(
     String routeName, {
     Object? args,